backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pymongo[srv]==4.6.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic[email]==2.6.3
aiofiles==23.2.1
cryptography>=41.0.0